+++ src/Plaid/Entity/EntityWatchlistScreeningSearchTerms.cs
@@ -51,5 +51,5 @@ public record EntityWatchlistScreeningSearchTerms
 	/// <para>The current version of the search terms. Starts at <c>1</c> and increments with each edit to <c>search_terms</c>.</para>
 	/// </summary>
 	[JsonPropertyName("version")]
-	public decimal Version { get; init; } = default!;
+	public int Version { get; init; } = default!;
 }
