+++ src/Plaid/Entity/ScreeningHitAnalysis.cs
@@ -33,5 +33,5 @@ public record ScreeningHitAnalysis
 	/// <para>The version of the screening's <c>search_terms</c> that were compared when the screening hit was added. screening hits are immutable once they have been reviewed. If changes are detected due to updates to the screening's <c>search_terms</c>, the associated program, or the list's source data prior to review, the screening hit will be updated to reflect those changes.</para>
 	/// </summary>
 	[JsonPropertyName("search_terms_version")]
-	public decimal SearchTermsVersion { get; init; } = default!;
+	public int SearchTermsVersion { get; init; } = default!;
 }
